% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezcor.R
\name{ezcor_batch}
\alias{ezcor_batch}
\title{Run correlation between two variables in a batch mode and support group by a variable}
\usage{
ezcor_batch(
  data,
  var1,
  var2,
  split = FALSE,
  split_var = NULL,
  cor_method = "pearson",
  adjust_method = "none",
  use = "complete",
  sig_label = TRUE,
  parallel = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data}{a `data.frame` containing variables}

\item{var1}{a `character`, the first variable in correlation}

\item{var2}{a `vector` containing variables in a batch mode}

\item{split}{whether perform correlation grouped by a variable, default is 'FALSE'}

\item{split_var}{a `character`, the group variable}

\item{cor_method}{method="pearson" is the default value. The alternatives to be passed to cor are "spearman" and "kendall"}

\item{adjust_method}{What adjustment for multiple tests should be used? ("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none")}

\item{use}{use="pairwise" will do pairwise deletion of cases. use="complete" will select just complete cases}

\item{sig_label}{whether add symbal of significance. P < 0.001,"***"; P < 0.01,"\\**"; P < 0.05,"*"; P >=0.05,""}

\item{parallel}{if `TRUE`, do parallel computation by **furrr** package.}

\item{verbose}{if `TRUE`, print extra info. If `parallel` is `TRUE`,
set `verbose` to `FALSE` may speed up.}
}
\value{
a `data.frame`
}
\description{
Run correlation between two variables in a batch mode and support group by a variable
}
\examples{
data("exprSet", package = "ezcor", envir = environment())
target_variable = colnames(exprSet)[3]
genelist <- colnames(exprSet)
genelist <-  setdiff(genelist,c("patient","tissue",target_variable))
res <- ezcor::ezcor_batch(data = exprSet,
                          var1 = target_variable,
                          var2 = genelist,
                          split = TRUE,
                          split_var = "tissue",
                          cor_method = "pearson",
                          adjust_method = "none",
                          use = "complete",
                          sig_label = TRUE,
                          parallel = FALSE,
                          verbose = TRUE)
}
\author{
Yi Xiong, Shixiang Wang
}
