% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezcor_bicor.R
\name{ezcor_bicor}
\alias{ezcor_bicor}
\title{Run Biweight Midcorrelation}
\usage{
ezcor_bicor(data, var1, var2, split = FALSE, split_var, ...)
}
\arguments{
\item{data}{a `data.frame` containing variables}

\item{var1}{a `character`, the first variable in correlation}

\item{var2}{a `character`, the second variable in correlation}

\item{split}{`logic value`,whether perform correlation grouped by a variable, default is 'FALSE'}

\item{split_var}{a `character`, the group variable}

\item{...}{other arguments passed to methods}
}
\value{
a `data.frame`
}
\description{
Run Biweight Midcorrelation
}
\examples{
data("exprSet", package = "ezcor", envir = environment())
g1 <- colnames(exprSet)[3]
g2 <- colnames(exprSet)[4]
genelist <- colnames(exprSet)
genelist <-  setdiff(genelist,c("patient","tissue",g1))
res <- ezcor::ezcor_bicor(data = exprSet,g1,g2)


}
\seealso{
\link{WGCNA::bicor()} which this function wraps.
}
\author{
Yi Xiong
}
