% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezcor_bicor.R
\name{ezcor_bicor_batch}
\alias{ezcor_bicor_batch}
\title{Run Biweight Midcorrelation in a batch mode}
\usage{
ezcor_bicor_batch(data, var1, var2, split = FALSE, split_var, ...)
}
\arguments{
\item{data}{a `data.frame` containing variables}

\item{var1}{a `character`, the first variable in correlation}

\item{var2}{a `vector` containing variables in a batch mode}

\item{split}{`logic value`, whether perform correlation grouped by a variable, default is 'FALSE'}

\item{split_var}{a `character`, the group variable}
}
\value{
a `data.frame`
}
\description{
Run Biweight Midcorrelation in a batch mode
}
\seealso{
[WGCNA::bicor()] which this function wraps.
}
\author{
Yi Xiong
}
