% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezcor_parallel.R
\name{ezcor_parallel}
\alias{ezcor_parallel}
\title{Run correlation between two variables in a batch mode and support group by a variable}
\usage{
ezcor_parallel(
  data,
  var1,
  var2,
  split = FALSE,
  split_var = NULL,
  Cor_method = "pearson",
  adjust_method = "none",
  sig_label = TRUE,
  parallel = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data}{a `data.frame` containing variables}

\item{var1}{a `character`, the first variable in correlation}

\item{var2}{a `vector` containing variables in a batch mode}

\item{split}{whether perform correlation grouped by a variable, default is 'FALSE'}

\item{split_var}{a `character`, the group variable}

\item{Cor_method}{method="pearson" is the default value. The alternatives to be passed to cor are "spearman" and "kendall"}

\item{adjust_method}{What adjustment for multiple tests should be used? ("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none")}

\item{sig_label}{whether add symbal of significance. P < 0.001,"***"; P < 0.01,"\\**"; P < 0.05,"*"; P >=0.05,""}

\item{parallel}{if `TRUE`, do parallel computation by **furrr** package.}

\item{verbose}{if `TRUE`, print extra info. If `parallel` is `TRUE`,
set `verbose` to `FALSE` may speed up.}
}
\value{
a `data.frame`
}
\description{
Run correlation between two variables in a batch mode and support group by a variable
}
\author{
Yi Xiong, Shixiang Wang
}
