% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezcor_partial_cor.R
\name{ezcor_partial_cor}
\alias{ezcor_partial_cor}
\title{Run partial correlation}
\usage{
ezcor_partial_cor(
  data = NULL,
  split = FALSE,
  split_var = NULL,
  var1 = NULL,
  var2 = NULL,
  var3 = NULL,
  cor_method = "pearson",
  sig_label = TRUE,
  ...
)
}
\arguments{
\item{data}{a `data.frame` containing variables}

\item{split}{whether perform correlation grouped by a variable, default is 'FALSE'}

\item{split_var}{a `character`, the group variable}

\item{var1}{a `character`, the first variable in correlation}

\item{var2}{a `character`, the second variable in correlation}

\item{var3}{a `character` or `character vector`, the third variable in correlation}

\item{cor_method}{method="pearson" is the default value. The alternatives to be passed to cor are "spearman" and "kendall"}

\item{sig_label}{whether add symbal of significance. P < 0.001,"***"; P < 0.01,"**"; P < 0.05,"*"; P >=0.05,""}

\item{...}{other arguments passed to methods}
}
\value{
a `data.frame`
}
\description{
Run partial correlation
}
\examples{
data("exprSet", package = "ezcor", envir = environment())
g1 <- colnames(exprSet)[3]
g2 <- colnames(exprSet)[4]
g3 <- colnames(exprSet)[c(5,6)]
res <- ezcor::ezcor_partial_cor(data = exprSet,
                                var1 = g1,
                                var2 = g2,
                                var3 = g3,
                                split = T,
                                split_var = "tissue",
                                cor_method = "pearson",
                                sig_label = TRUE)


}
\seealso{
\link{ppcor::pcor.test()} which this function wraps.
}
\author{
Yi Xiong
}
